Refactor integration tests: remove centralized CRUD, batch, query, streaming, and error handling tests; add KISS tests for PocketBase integration.
